fix: Correct snackbar UI and make it hide as expected
When a post was deleted right after its creation (in fact it was needed to wait
about 6 seconds) the deleted post snackbar would stole focus from the Add Post
page which the browser would be returned to. That was leading to snackbar's
hidden action button to appear as a traslucid box, and preventing the snackbar
to hide, and stoling intended focus from Add Post's element.